ci: run the census-instrument tests off the censuses job, not the Test shards
The four run-backed tests in `conformance_census_instrument` (#2063) each
normalize the whole spec corpus, costing 60-271s apiece. The module was only
enrolled in `ORACLE_EXCLUDE`, not in `CENSUS_FILTER`, so it ran in full in the
6-way `Test` shards. On run 31973592268 both 60s+ tests hashed into `Test (1/6)`,
which ran 277.6s while its other 1,885 tests finished in ~72s — three quarters of
the shard's wall was two tests. `--partition hash:K/N` balances shard counts, not
cost, and a shard's wall cannot fall below its slowest single test, so re-sharding
cannot fix it.

Split the four full-census tests into a new `conformance_census_runs` module and
enroll it in the census carve-out exactly as the four existing census modules are:
`#[path]`-included in the soak driver, named in `CENSUS_FILTER` (negated from the
shards and `test-oracle`) and selected by the censuses job's un-armed step
(`run_census` refuses with a seam oracle armed). It is named in `ORACLE_EXCLUDE`
too, because it imports `conformance::census`.

The pure tests and the `#[cfg(feature = "benchmark")]` refusal test stay in
`conformance_census_instrument` in the shards, both cheap — the refusal test needs
the `ferro-benchmark` binary the soak driver does not build, so it must. The moved
tests are otherwise unchanged.
