+++ tests/it/main.rs
@@ -116,6 +116,7 @@ mod comprehensive_external_tests;
 mod con_conversion_audit;
 mod confluence_self_check_tests;
 mod conformance_census_instrument;
+mod conformance_census_runs;
 mod conformance_summary_generated;
 mod convert_gff_library_parity;
 mod convert_tests;
